Argument parser change
Now, you can supply the path of the file containing the characters to write through the command line. In case you don't supply anything, it defaults to char.txt (Note: The program will assume that char.txt is in the current working directory. In case it isn't, you must supply the path on your own)